==> scripts/1.data_wrangling_methods.Rmd <==
---
title: "Data Wrangling "
author: "Indra Boving & Joe Celebrezze"
date: "8/16/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(here)
here = here::here
library(corrplot)
library(janitor)
select = dplyr::select
col3 = hcl.colors(10, "YlOrRd", rev = TRUE) #color scale to use for consistency 
```

# SEKI DATA: 

```{r}
seki_all_methods <- read_csv(here("processed-data", "seki_flam_data_all.csv"), show_col_types = FALSE) %>% 
  clean_names() %>% 
  mutate(lfm = lfm_n_as_imputed) %>% 
  select(2:49, max_mpa_sample, -ttms, -bottle_wt, lfm) %>%  #drop columns we don't want
  select(-model_spp_gr, -prop_ignite, -ros) %>% 
  mutate(location = "seki") %>% 
  mutate(gdw_gfw = dry_wt/fresh_wt) %>%
  #mutate(gww_gdw = lfm/100) %>%
  mutate(dw_flam_sample = (sample_wt*(1/(1+lfm/100))))

#This is a good way to look at the dataset: 
#dlookr::diagnose(seki_all_methods) 
```

# LOCAL DATA: 

```{r}
local_all_methods <- read_csv(here("raw-data", "local_flam_data_all.csv"), show_col_types = FALSE) %>% clean_names() %>%
  mutate(location = "local") %>% 
  mutate(ignition_no_double = 1) %>% 
  dplyr::mutate(gdw_gfw = dry_wt/fresh_wt) %>%
  #dplyr::mutate(gww_gdw = lfm/100) %>%
  #dplyr::mutate(gdw_gww = dry_wt/water_wt) %>%
  dplyr::mutate(dw_flam_sample = (sample_wt*(1/(1+lfm/100)))) %>% 
  dplyr::mutate(ww_flam_sample = sample_wt - dw_flam_sample) %>% 
  select(-gww_gdw, -flam_index) # Removing unnecessary column, gww_gdw, as well as the flam_index (since it is not correctly calculated)

#This is a good way to look at the dataset: 
#dlookr::diagnose(local_all_methods)
```

Old DW formula: (sample_wt * 100)/lfm

# COMBINED:
```{r}
all_methods_long <- rbind(seki_all_methods, local_all_methods) %>% 
  unite(individual_round, c(individual, round), sep = "_", remove = FALSE)

all_methods_epi <- all_methods_long %>% 
  filter(model == "EPI")

all_methods_hp <- all_methods_long %>% 
  filter(model == "HP")
```

## Flam. index (Essaghi et. al. 2017)
- although we do not believe this is going to be used heavily in the analyses as it slightly extends the scope of this paper beyond what we deem to be the acceptable limits for the scope, we include it for exploratory/extra analyses 

This is the equation used for the flam. index:
\start{equation}
flam_index = x \times (\frac{max(tti) - tti}{min(tti) + tti})^{\frac{fh}{fh + max(fh)}^2}
\end{equation}
where max(tti) and max(fh) were determined for each method prior to calculating and where x is some arbitrary coefficient so that the maximum value off the flammability index (i.e., tti = 0, fh = max(fh)) is equal to 20
```{r}
# Epiradiator
all_methods_epi <- all_methods_epi %>% 
  mutate(flam_index = 10.936*((86-tti)/(7+tti))^((fh)/(fh + 27))^2)

# Hot plate
all_methods_hp <- all_methods_hp %>% 
  mutate(flam_index = 9.306*((192-tti)/(9+tti))^((fh)/(fh + 54))^2)
```

```{r}
#Make a dataset that only has the rounds where we have paired EPI and HP samples
paired_methods_wide <- merge(all_methods_epi, all_methods_hp, by.x = "individual_round", by.y = "individual_round", suffixes = c("_epi", "_hp"))
```

# Slight issues
- Adding in water weight (sample weight - dry weight)
```{r}
paired_methods_wide <- paired_methods_wide %>% 
  mutate(ww_flam_sample_epi = sample_wt_epi - dw_flam_sample_epi) %>% 
  mutate(ww_flam_sample_hp = sample_wt_hp - dw_flam_sample_hp)
```

- All MPa values should be negative
```{r}
paired_methods_wide <- paired_methods_wide %>% 
  mutate(mpa_epi = case_when(mpa_epi > 0 ~ mpa_epi*-1, mpa_epi < 0 ~ mpa_epi)) %>% 
  mutate(mpa_hp = case_when(mpa_hp > 0 ~ mpa_hp*-1))
```

- NA Value for HP > ADFA > TTI, but ignition present
Imputing median TTI for ADFA, using only natural ignitions (not including manual ignitions)
```{r}
median.adfa.no.manuals <- paired_methods_wide %>% 
  filter(spp_hp == "ADFA", ignition_hp == 1)
median(median.adfa.no.manuals$tti_hp, na.rm = T) #96

paired_methods_wide <- paired_methods_wide %>% 
  mutate(tti_hp = ifelse(is.na(tti_hp),
                            96,
                            tti_hp))
```

# Main Dataframe (Long)
```{r}
paired_methods_epi <- paired_methods_wide %>% 
  select(1:48) %>% 
  rename_with(~str_remove(., '_epi'))

paired_methods_hp <- paired_methods_wide %>% 
  select(1, 49:95) %>% 
  rename_with(~str_remove(., '_hp'))

paired_methods_long <- bind_rows(paired_methods_epi, paired_methods_hp)
```


# Removing Points
Below, I have (1) removed species that have N = 6 or less (QUKE, ABCO, PIJE) and (2) taken out manual ignitions. I am going to keep one dataset (bothm_bothl_manual_ignitions_in.csv) which has the manual ignitions in case we want to use them at some point.
```{r}
paired_methods_long <- paired_methods_long %>% 
  filter(spp != "QUKE") %>% 
  filter(spp != "ABCO") %>% 
  filter(spp != "PIJE") #taking out these species
paired_methods_long2 <- paired_methods_long %>% 
  filter(ignition != 2) #taking out manual ignitions

paired_methods_wide <- paired_methods_wide %>% 
  filter(spp_epi != "QUKE") %>% 
  filter(spp_epi != "ABCO") %>% 
  filter(spp_epi != "PIJE") #taking out these species
paired_methods_wide2 <- paired_methods_wide %>% 
  filter(ignition_epi != 2)
```

```{r}
epi.sum <- paired_methods_long %>% 
  filter(model == "EPI") %>% 
  select(mpa, lfm, start_temp, tti, fd, gd, fh, ignition, spp) %>% 
  group_by(spp) %>% 
  summarise(mean.mpa = mean(mpa), mean.lfm = mean(lfm), mean.start.temp = mean(start_temp), mean.tti = mean(tti), mean.fd = mean(fd), mean.gd = mean(gd), mean.fh = mean(fh),total.prop.ignite = mean(ignition), n = n())
epi.sum$model <- "EPI"

hp.sum <- paired_methods_long %>% 
  filter(model == "HP") %>% 
  select(mpa, lfm, start_temp, tti, fd, gd, fh, ignition, spp) %>% 
  group_by(spp) %>% 
  summarise(mean.mpa = mean(mpa), mean.lfm = mean(lfm), mean.start.temp = mean(start_temp), mean.tti = mean(tti), mean.fd = mean(fd), mean.gd = mean(gd), mean.fh = mean(fh), total.prop.ignite = mean(ignition), n = n())
hp.sum$model <- "HP"

summary.df <- rbind(epi.sum, hp.sum)
```

# Writing csv. files
```{r, include = FALSE}
write_csv(paired_methods_long2, here("processed-data", "bothmethods_bothlocations_long.csv"))

write_csv(paired_methods_long, here("processed-data", "bothm_bothl_manual_ignitions_in.csv"))

write_csv(paired_methods_wide2, here("processed-data", "bothmethods_bothlocations_wide.csv"))
```
