==> scripts/2.literature_review.Rmd <==
---
title: "Literature Review"
author: "Joe Celebrezze"
date: '2022-08-26'
output: html_document
---

Set-up
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
select = dplyr::select
library(here)
here = here::here
library(ggpubr) #to put graphs side by side
library(sf) #for dealing with shapefiles
library(gcookbook) #for continuous color palette
library(kableExtra) #for table A1

data <- read_csv(here('raw-data', 'lit.review.csv'))[1:158,]
location.data <- read_csv(here('raw-data', 'lit.review.locations.csv'))[1:184,]
drydown.data <- read_csv(here('raw-data', 'lit.review.drydown.csv'))[1:157,] # remove last row since NA for drydown.method
```

Global ignitions datasets
NOTE: these datasets are *very large*. They are too large to push to the GitHub in their raw form, so *if they are downloaded, they must be deleted from the remote prior to committing and pushing changes to GitHub*. If they are left on the remote, an error will pop up saying that the file is too large. Also, (as if that first part wasn't already annoying enough) these urls will expire on December 30, 2022 and then I will need to request them from Global Fire Atlas again to get new urls. These complications unfortunately make this code difficult to distribute.
```{r}
# Ignitions (2016) Dataset from global fire atlas
temp <- tempfile()
url <- 'https://daac.ornl.gov/orders/59d386f26ee3bdb165507bd6ff4134df/CMS_Global_Fire_Atlas/data/Global_fire_atlas_V1_ignitions_2016.zip' # Note: this URL may expire after 30 days (in March)
download.file(url, temp)
fire.map2016 <- unzip(zipfile = temp)
fire.map2016 <- sf::read_sf('Global_fire_atlas_V1_ignitions_2016.shp')

# Ignitions (2015) Dataset from global fire atlas
temp <- tempfile()
url <- 'https://daac.ornl.gov/orders/59d386f26ee3bdb165507bd6ff4134df/CMS_Global_Fire_Atlas/data/Global_fire_atlas_V1_ignitions_2015.zip' # Note: this URL may expire after 30 days (in April)
download.file(url, temp)
fire.map2015 <- unzip(zipfile = temp)
fire.map2015 <- sf::read_sf('Global_fire_atlas_V1_ignitions_2015.shp')

# Combining two datasets above
fire.map <- rbind(fire.map2016, fire.map2015)
```

# Bar Graph
Setting up dataframe
```{r}
group <- c('Epiradiator', 'Wind Tunnel',	'Muffle Furnace',	'Mass Loss\n Calorimeter',	'Xylene Lattice',	'Fire Bench',	'Grill',	'Oxygen Bomb',	'Steel Ring',	'Tray',	'Flat-Flame Burner',	'FPA Calorimeter',	'Cone Calorimeter',	'Direct Flame',	'Other',	'Intermediate\n Scale Calorimeter')
value <- c(rep(NaN, 16))
data.summary <- data.frame(group, value)

for(i in c(1:16)) {
  data.summary$value[i] <- sum(data[,i+11])/nrow(data)
  data.summary$count[i] <- sum(data[,i+11])
}

data.summary$group <- factor(data.summary$group, levels = c('Epiradiator', 'Grill', 'Muffle Furnace', 'Cone Calorimeter', 'Mass Loss\n Calorimeter', 'Oxygen Bomb','Steel Ring', 'Wind Tunnel', 'Fire Bench', 'Xylene Lattice', 'Direct Flame', 'Tray', 'FPA Calorimeter', 'Flat-Flame Burner',	'Intermediate\n Scale Calorimeter', 'Other'))
```

Visualization
```{r}
ggplot(data.summary, aes(x=group, y=value, fill=group)) +
  geom_bar(stat="identity", width=1, color="white") +
  labs(y = 'Proportion of Experiments Using Method') +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title.x = element_blank(),
        axis.title.y = element_text(face = 'bold'),
        axis.text.x = element_text(face = 'bold', angle = 90, hjust = 1,vjust = 0.4, size = 12))
```

```{r}
ggplot(data.summary, aes(x=group, y=count, fill=group)) +
  geom_bar(stat="identity", width=1, color="white") +
  labs(y = 'Number of Experiments Using Method') +
  theme_bw() +
  ylim(c(0, 46.5)) +
  annotate('text', x = 1, y = 44, label = 'b', size = 14, fontface = 'bold') +
  theme(legend.position = 'none',
        axis.title.x = element_blank(),
        axis.title.y = element_text(face = 'bold', size = 14),
        axis.text.x = element_text(face = 'bold', angle = 90, hjust = 1,vjust = 0.4, size = 20))

ggsave(here('figures/Fig2b.lit.review.barplot.jpg'), height = 7, width = 12.64)
```

# Comparing Weights
## Epiradiator Only
```{r}
data %>% filter(model == 'EPI') %>% 
ggplot(aes(x = sample.weight)) +
  labs(x = "Sample Weight (g)", y = "Density", title = "Epiradiator Flammability Tests") +
  geom_density() +
  scale_x_continuous(trans = 'log10') +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold'))

ggsave(here('figures/lit.review.epi.weights.jpg'), height = 5, width = 7.5)
```

## All Models
```{r}
ggplot(data = data, aes(x = sample.weight)) +
  labs(x = "Sample Weight (g)", y = "Density") +
  geom_density(size = 1.2) +
  scale_x_continuous(trans = 'log10', breaks = c(0.1, 1, 10, 100, 1000), labels = c(0.1, 1, 10, 100, 1000)) +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 14),
        axis.text = element_text(size = 12))

ggsave(here('figures/lit.review.all.methods.weights.jpg'), height = 5, width = 7.5)
```

## All Models & Epiradiator Specifically
```{r}
data.epi <- data %>% 
  filter(model == 'EPI')
# Plot
plot.s1 <- ggplot() +
  labs(x = "Sample Weight (g)", y = "Kernel Probability Density") +
  geom_density(data = data, aes(x = sample.weight), size = 1.2) +
  geom_density(data = data.epi, aes(x = sample.weight), size = 1.2, alpha = 0.3, color = 'grey70') +
  scale_x_continuous(trans = 'log10', breaks = c(0.1, 1, 10, 100, 1000), labels = c(0.1, 1, 10, 100, 1000)) +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 14),
        axis.text = element_text(size = 12))
plot.s1
# Legend
legend.s1 <- get_legend( 
  data %>%   
  mutate(legend.tag = if_else(model == 'EPI', 'Epiradiator', 'All Models')) %>% 
  ggplot(aes(x = sample.weight, y = year, color = legend.tag)) +
    geom_smooth(size = 1.2, se = F) +
    labs(color = ' ') +
    scale_color_manual(values = c('black', 'grey70')) +
    theme_bw() +
    theme(legend.title = element_text(face = 'bold', size = 14),
          legend.text = element_text(size = 12)))
legend.s1
# Arranging
main.plot.s1 <- cowplot::plot_grid(plot.s1, legend.s1, ncol = 2, rel_widths = c(4,1))
main.plot.s1
# Saving
ggsave(here('figures/FigS1.lit.review.weights.jpg'), height = 5, width = 9)
```

# Map!
Cleaning up dataframes
```{r}
location.data.clean <- location.data %>% 
  select(-`location notes`, -study) %>% 
  na.omit()

# Focusing on Class F and G fires (defined by the National Wildfire Coordinating Group as fires 1000 acres or larger)
large.fire.map <- fire.map %>% 
  filter(size > 4.04686)

# Focusing on Class G fires (defined by the National Wildfire Coordinating Group as fires 5000 acres or larger)
v.large.fire.map <- fire.map %>% 
  filter(size > 20.23428)

# California map
states <- map_data("state")
states <- states %>% 
  filter(region == 'california')

# Classifying fire sizes
fire.map <- fire.map %>% 
  mutate(size.class = case_when(
    size > 0 & size <= 1 ~ '0-1km',
    size > 1 & size <= 2.5 ~  '1-2.5km',
    size > 2.5 & size <= 5 ~ '2.5-5km',
    size > 5 & size <= 20 ~ '5-20km',
    size > 20 & size <= 100 ~ '20-100km',
    size > 100 & size <= 1000 ~ '100-1000km',
    size > 1000 ~ '>1000km'
  )) %>% 
  mutate(size.class = as.factor(size.class))
fire.map$size.class <- factor(fire.map$size.class, levels = c('>1000km', '100-1000km','20-100km', '5-20km', '2.5-5km','1-2.5km','0-1km'))
levels(fire.map$size.class)
```

## All Ignitions
```{r}
world <- map_data("world")
ggplot(data = world) +
  geom_map(data = world, map = world, aes(x = long, y = lat, map_id = region), fill = 'white') +
  geom_map(data = states, map = states, aes(x = long, y = lat, map_id = region), fill = 'gray85') +
  ylim(-55,90) +
  geom_point(data = fire.map, aes(x = longitude, y = latitude, color = size.class), alpha = 0.035, size = 0.008) +
  labs(color = "Fire Size (km2)") +
  scale_color_manual(values = c('#2d0300', '#663417', '#ac5011', '#df6100', '#e9802b', '#f5b05d', '#f5f05d')) +
  #scale_color_manual(values = c('y))
  #scale_color_gradient2(low = '#FFF06D', mid = '#AF5A26', high = '#840900', trans = 'log', midpoint = 3.65, breaks = c(1, 5, 25, 125, 625, 3000, 9000)) +
  geom_point(data = location.data.clean, aes(x = longitude, y = latitude), color = 'black', size = 1.25, alpha = 0.7) +
  annotate('text', x = -172, y = 85, label = 'a', size = 14, fontface = 'bold')+
  theme_void() +
  guides(color = guide_legend(override.aes = list(size=5, alpha = 0.8))) +
  theme(panel.background = element_rect(fill = '#0AA3CF'),
        legend.position = c(0.13, 0.2),
        legend.title = element_text(face = 'bold', size = 14),
        legend.text = element_text (size = 11))

ggsave(here('figures/Fig2a.lit.review.locations.all.ignitions.jpg'), height = 7, width = 12.64)
```

## Ignitions for Large Fires (Classes F and G)
```{r}
ggplot(data = world) +
  geom_map(data = world, map = world, aes(x = long, y = lat, map_id = region), fill = 'white') +
  ylim(-55,90) +
  geom_point(data = large.fire.map, aes(x = longitude, y = latitude, color = size), alpha = 0.05, size = 0.05) +
  labs(color = "Fire Size (km2)") +
  scale_color_gradient(low = '#EFA00B', high = '#CA3C25', trans = 'log', breaks = c(5, 25, 125, 625, 3000, 9000)) +
  geom_point(data = location.data.clean, aes(x = longitude, y = latitude), color = 'black', size = 1.25, alpha = 0.7) +
  theme_void() +
  theme(panel.background = element_rect(fill = '#0AA3CF'),
        legend.position = c(0.13, 0.2),
        legend.title = element_text(face = 'bold', size = 14),
        legend.text = element_text (size = 12))

ggsave(here('figures/extra-figures/lit.review.locations.FG.ignitions.jpg'), height = 7, width = 12.64)
```

## Ignitions for Large Fires (Class G)
```{r}
ggplot(data = world) +
  geom_map(data = world, map = world, aes(x = long, y = lat, map_id = region), fill = 'white') +
  ylim(-55,90) +
  geom_point(data = v.large.fire.map, aes(x = longitude, y = latitude, color = size), alpha = 0.05, size = 0.05) +
  labs(color = "Fire Size (km2)") +
  scale_color_gradient(low = '#EFA00B', high = '#CA3C25', trans = 'log', breaks = c(25, 125, 625, 3000, 9000)) +
  geom_point(data = location.data.clean, aes(x = longitude, y = latitude), color = 'black', size = 1.25, alpha = 0.7) +
  theme_void() + 
  theme(panel.background = element_rect(fill = '#0AA3CF'),
        legend.position = c(0.13, 0.2),
        legend.title = element_text(face = 'bold', size = 14),
        legend.text = element_text (size = 12))

ggsave(here('figures/extra-figures/lit.review.locations.G.ignitions.jpg'), height = 7, width = 12.64)
```

## California Map
This map includes our sites!
```{r}
# Coordinates for our sites (using site 1 at each location, as site 1 and site 2 are rather close to each other in both locations, so mapping both sites at both locations doesn't make much sense aesthetically)
site.labels <- c('Chaparral', 'Sierra Nevadas')
site.latitude <- c(34.4613, 36.7209)
site.longitude <- c(-119.6934, -118.9708)
site.coordinates <- data.frame(site.labels, site.latitude, site.longitude)

# Removing point in Nevada
location.data.ca <- location.data.clean %>% 
  filter(latitude != 40.28)

ggplot(data = states) +
  geom_map(data = states, map = states, aes(x = long, y = lat, map_id = region), fill = 'gray85') +
  xlim(c(-124.4, -114)) +
  ylim(c(32, 42.03)) +
  geom_point(data = fire.map, aes(x = longitude, y = latitude, color = size.class), alpha = 0.25, size = 0.5) +
  scale_color_manual(values = c('#2d0300', '#663417', '#ac5011', '#df6100', '#e9802b', '#f5b05d', '#f5f05d')) +
  geom_point(data = location.data.ca, aes(x = longitude, y = latitude), color = 'black', size = 4, alpha = 0.7) +
  geom_point(data = site.coordinates, aes(x = site.longitude, y = site.latitude), color = '#f78c45', size = 6, shape = 18) +
  geom_text(data = site.coordinates, aes(x = site.longitude, y = site.latitude + 0.45, label = site.labels), size = 8) +
  theme_void() +
  theme(panel.background = element_rect(fill = '#0AA3CF'),
        legend.position = 'none')

ggsave(here('figures/Fig2a.2.california.inset.jpg'), height = 6, width = 6)
```

# Drydown Types
```{r}
drydown.data.clean <- drydown.data %>% 
  group_by(drydown.method) %>% 
  summarise(counts = n())

drydown.data.clean$drydown.method <- factor(drydown.data.clean$drydown.method, levels = c('no', 'oven', 'air', 'fresh.oven', 'lab.drydown', 'dead.live', 'field.drought', 'fresh.air.oven', 'air.oven', 'fresh.air', 'potted.plants', 'other'))

ggplot(data = drydown.data.clean, aes(x = drydown.method, y = counts)) +
  geom_bar(stat = 'identity', fill = 'firebrick') +
  labs(x = 'Drydown Method', y = 'Count') +
  geom_text(aes(label = counts), vjust = -0.3, color = 'firebrick', fontface = 'bold') + 
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, size = 14),
        axis.title = element_text(size = 16, face = 'bold'))
```

# Table A1
Descriptions of all methods used and studies that used the given method, included in the appendix
```{r}
# Methods
methods <- c('Epiradiator', 'Grill', 'Muffle Furnace', 'Cone Calorimeter', 'Mass Loss Calorimeter', 'Oxygen Bomb','Steel Ring', 'Wind Tunnel', 'Fire Bench', 'Xylene Lattice', 'Direct Flame', 'Tray', 'FPA Calorimeter', 'Flat-Flame Burner',	'Intermediate Scale Calorimeter', 'Other')

# Descriptions
epi.description <- 'This method is defined by a Quartzalliance epiradiator as the primary heating source. Suspension methods (i.e., wire mesh), inclusion of a pilot flame, sample weights, heat flux, starting temperature, locations of thermocouples vary across different studies. Studies investigating bark flammability (e.g., [63]) have the surface of the epiradiator in a vertical orientation; however, the epiradiator is typically in a horizontal orientation, as the majority of studies using this method investigate either live vegetation or litter layers.'
grill.description <- 'The grill apparatus was designed and described in detail in [9] In this case, it involves a 85x60cm metal barrel cut in half, a metal grill placed in between each half with propane-powered burners below the grill as heat source, a blow torch as an ignition source, and removable wind protection in an attempt to control the environment between burns. In some other cases, a manufactured BBQ grill was used in place of the barrel-based grill, as they are more easily accessible.'
mf.description <- 'This method consists of placing a sample inside a muffle furnace and measuring flammability parameters such as time to ignition. Samples are often individual leaf specimens. Temperature of the muffle furnace varied between studies from 400-700°C. The door of the muffle furnace is kept open during tests to allow for the measurement of flammability metrics.'
cc.description <- 'This calorimetry method measures heat release rate by measuring the consumption of oxygen resulting from combustion. It consists of a conical heater with a fixed heat flux to heat samples. The ignition source varies between tests with the more common ones including a spark ignitor and a pilot flame. Tests are typically performed in accordance with ATSM E 1354 standards [107]. ** In some cases, the iCone calorimeter (Fire Testing Technology, East Grinstead, UK) was used. This is an advanced, automatic cone calorimeter designed to test a variety of flammability parameters.'
mlc.description <- 'This is a specific type of calorimeter, manufactured by Fire Testing Technology Limited (FTT®), which is very similar to a cone calorimeter, but with adjustments to specialize in full fires and to measure smaller changes in mass loss. It also looks at heat release and heat flux. The MLC involves a conical heater, a 500g-capacity load cell, a spark ignition source, a chimney with a thermopile. Tests were most commonly conducted at 50 kW/m2.'
ob.description <- 'This is a calorimetry method commonly used to measure gross energy content and heat of combustion. It operates by having a reaction occur in a container (referred to as the ‘bomb’) with a fixed volume, so that changes in temperature can be attributed to the energy flux of the reaction rather to any changes in volume. In this review, it was always paired with other methods.'
sr.description <- 'This is similar to the ‘tray’ method, but it involves a more sophisticated setup to maximize airflow and can incorporate both leaf litter and/or vegetation samples. It involves a steel mesh ring with samples loaded into it. Often, it also has a fireproof base, such as one made from a hardiflex cement fiber board. Ignition sources vary, including a paper towel soaked in isopropyl alcohol and a cotton ball soaked in ethanol.'
wt.description <- 'This group consists of any method that has a wind tunnel directing wind towards the samples emulating winds that play a key role during wildfires. Sample size and composition varies across the studies, from looking at full, reconstructed shrubs and litter layers to looking at 4g samples. A heat source, such as an infrared panel, is often utilized and usually located above a sample. An ignition source is typically present and varies by study, with the most common sources including a firebrand and a coiled wire igniter.'
fb.description <- 'This method involves a large flat surface in which large plant samples or, more commonly, litter beds are tested for flammability. Ignition sources vary for the different studies and include ‘firebrands’ (usually a small block of wood, ignited by means of an epiradiator), cotton wicks soaked in a flammable liquid such as ethanol, and a propane torch.'
xl.description <- 'This is another commonly used method to test litter bed flammability. It involves placing litter beds atop a lattice of cotton strings soaked in xylene, a flammable liquid, and then igniting the lattice of strings from one or multiple points leading to rapid ignition and combustion of the litter bed.'
direct.flame.description <- 'Samples are directly ignited by an open flame without any other heat source present. The open flame varies across studies from a bunsen burner to a larger U-shaped gas burner'
tray.description <- 'This is yet another method to test litter bed flammability. It involves placing litter beds in a metal tray with limited airflow from the sides or bottom. An ignition source, typically a firebrand, is placed in the center or edge of the litter bed to cause ignition and combustion of the litter bed.'
fpa.description <-  'This method is similar to the cone calorimeter, but the combustion chamber allows for a more controlled environment with less stochasticity in gas flow rate and composition.'
ffb.description <- 'This method was designed to emulate a moving flame front. Small samples (typically, individual leaves) are suspended and a flat flame burner moves towards the sample by route of a small motor. Sometimes, a 6000W radiative panel (Omega) is also included as a radiative heat source.'
isc.description <- 'This is also similar to a cone calorimeter, but it is designed for larger scale samples, typically burning full plants. It consists of equipment to measure gaseous concentrations, a propane line burner to create a vertical ‘wall of flame’, three wooden walls lined with ceramic fibreboard, and a plant rack in the middle of the three wooden walls.'
other.description <- '(1) Litter was placed in a cubic basket which was then loaded into a recirculating air oven which sustained temperatures up to 250°C, while thermocouples placed in the middle of the litter took temperature readings.
(2) This is a specialized experiment investigating the propagation of flames from litter layers to shrub canopies. 20x20cm litter samples were dried and then ignited 10cm below a steel wire netting holding a variety of shrub samples representative of the shrublands of SE France to emulate litter-to-shrub fire spread
(3) Litter was placed in a cubic cage which was loaded atop a sand surface into a 1.09m-tall chamber made of vermiculite insulation board. Inside the chamber, the sample was heated using an IR lamp and a handheld spark generator was used as an ignition source;
(4) Leaf-level flammability was measured using a Federal Aviation Administration (FAA, USA) microcalorimeter using 10-15mg samples;
(5) Shoot-level flammability was measured using a horizontal 4kW radiant panel located 1cm beneath the shoot sample which was suspended by a specialized holder, with a heat flux sensor measuring heat flux and a scale beneath the sample holder measuring weight consumed during burning.
(6) Full-plant flammability was measured using potted plants. Aluminum discs were located at soil level and 50cm above soil level to estimate heat release. The plant was ignited by a burning cotton ball, soaked in 10mL of ethanol.'
descriptions <- c(epi.description, grill.description, mf.description, cc.description, mlc.description, ob.description, sr.description, wt.description, fb.description, xl.description, direct.flame.description, tray.description, fpa.description, ffb.description, isc.description, other.description)

epi.studies <- '[5,6],[16,17],[24],[30],[32],[45],[67-91],[15,18,31,33,34]*'
grill.studies <- '[11],[92-104]'
mf.studies <- '[8],[46],[55,56],[100],[105-111]'
cc.studies <- '[113]*,[114-121],[48,122]**'
mlc.studies <- '[9],[12,13],[47],[59],[67],[123-126]'
ob.studies <- '[6], [8,9], [12],[75],[79],[89],[93],[127,128]'
sr.studies <- '[129-137]'
wt.studies <- '[12,13],[50],[57],[138-142]'
fb.studies <- '[17],[54],[125],[143-148]'
xl.studies <- '[20],[149-154]'
direct.flame.studies <- '[19],[127,128],[132],[155,156]'
tray.studies <- '[24],[139],[157-160]'
fpa.studies <- '[161-164]'
ffb.studies <- '[7],[58],[165,166]'
isc.studies <- '[50],[116]'
other.studies <- '(1) [167];(2) [160];(3) [168];(4) [48];(5) [27];(6) [169]'
studies <- c(epi.studies, grill.studies, mf.studies, cc.studies, mlc.studies, ob.studies, sr.studies, wt.studies, fb.studies, xl.studies, direct.flame.studies, tray.studies, fpa.studies, ffb.studies, isc.studies, other.studies)

table.a1.df <- data.frame(methods, descriptions, studies)
```

```{r}
table.a1.df %>% 
  kable(format = 'html', escape = F, col.names = c('Method', 'Description', 'Studies'), caption = '* : the six studies identified with an asterisk used the laboratory benchtop drydown process described in our methods and in pressure-volume curve protocol [26]') %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  row_spec(c(0:16), color = 'black') %>% 
  column_spec(1, bold = T, border_left = T, border_right = T) %>%  
  column_spec(2:3, border_left = T, border_right = T) %>% 
  row_spec(c(2,4,6,8,10,12,14,16), background = '#F9F6EE') %>% 
  save_kable(here('figures/TableA1.jpg'))
```
